requirements.txt
Flask
python-dotenv
requests
PyJWT
fpdf2
openpyxl
pytz